move parse kpts func to utils
